Update OptiWin.bat
Changed the 2nd Win32Priority Separation value to 42 decimal or 2A hex
Changed the wording of the different menus to help better project what I want in this project.
Removed some of the placebo tweaks, theres still a bit to go. This application does help a lot and I am currently testing the tweaks that do not change a whole lot. (looking at you network tweaks)